fix(api/ui): fixed bugs from testing

+++ API/DTO/ProductNutrition/CreateProductRequest.cs
@@ -4,10 +4,10 @@ public class CreateProductRequest
 {
     public string Name { get; set; }
     public string CategoryName { get; set; }
-    public int CaloriesPer100Grams { get; set; }
-    public int ProteinPer100Grams { get; set; }
-    public int FatPer100Grams { get; set; }
-    public int CarbohydratesPer100Grams { get; set; }
+    public decimal CaloriesPer100Grams { get; set; }
+    public decimal ProteinPer100Grams { get; set; }
+    public decimal FatPer100Grams { get; set; }
+    public decimal CarbohydratesPer100Grams { get; set; }
 
     public IFormFile? File { get; set; }
 }
