+++ API/DTO/ProductNutrition/UpdateProductRequest.cs
@@ -10,13 +10,13 @@ public class UpdateProductRequest
     [Required]
     public string CategoryName { get; set; }
     [Required]
-    public int CaloriesPer100Grams { get; set; }
+    public decimal CaloriesPer100Grams { get; set; }
     [Required]
-    public int ProteinPer100Grams { get; set; }
+    public decimal ProteinPer100Grams { get; set; }
     [Required]
-    public int FatPer100Grams { get; set; }
+    public decimal FatPer100Grams { get; set; }
     [Required]
-    public int CarbohydratesPer100Grams { get; set; }
+    public decimal CarbohydratesPer100Grams { get; set; }
 
     public IFormFile? File { get; set; }
 }
