+++ API/DTO/ProductRecord/ProductRecordResponse.cs
@@ -4,9 +4,9 @@ public class ProductRecordResponse
 {
     public int Id { get; set; }
     public string Name { get; set; }
-    public double Grams { get; set; }
+    public decimal Grams { get; set; }
     public int Calories { get; set; }
-    public double Protein { get; set; }
-    public double Fat { get; set; }
-    public double Carbohydrates { get; set; }
+    public decimal Protein { get; set; }
+    public decimal Fat { get; set; }
+    public decimal Carbohydrates { get; set; }
 }
