+++ API/Entities/ProductNutrition.cs
@@ -5,10 +5,10 @@ public class ProductNutrition
     public int Id { get; set; }
     public int ProductNutritionCategoryId { get; set; }
     public string Name { get; set; }
-    public int CaloriesPer100Grams { get; set; }
-    public double ProteinPer100Grams { get; set; }
-    public double FatPer100Grams { get; set; }
-    public double CarbohydratesPer100Grams { get; set; }
+    public decimal CaloriesPer100Grams { get; set; }
+    public decimal ProteinPer100Grams { get; set; }
+    public decimal FatPer100Grams { get; set; }
+    public decimal CarbohydratesPer100Grams { get; set; }
 
     public string? ImageUrl { get; set; }
